rename api schemas
